requirements.txt
selenium
pandas
python-dotenv
openai
anthropic
gspread
oauth2client
openpyxl
